[style] chat styles 설정

+++ src/tools/instance.jsx
@@ -7,42 +7,47 @@ const instance = axios.create({
 
 // 요청 인터셉터 추가
 instance.interceptors.request.use(
-  async (config) => {
+  (config) => {
     // 요청이 전달되기 전 작업 수행
-    // console.log("인터셉터리퀘스트:", config);
     const token = localStorage.getItem("token");
     if (token) {
       config.headers["Authorization"] = token;
-      // console.log("1번째 토큰!", token);
     }
     return config;
   },
   (error) => {
-    // 요청 오류가 있는 경우 작업 수행
-    // console.log("인터셉터에러", error);
     return Promise.reject(error);
   }
 );
-// // 응답 인터셉터 추가
+
 instance.interceptors.response.use(
   async (response) => {
     // 응답 데이터가 있는 작업 수행
-    //console.log("인터셉터리스판스+++++++++++++++++:", response);
+    const { config, data } = response;
+    const prevRequest = config;
+
     if (response.status === 200 && response.data.code === 1) {
-      window.localStorage.removeItem("token");
-      window.localStorage.setItem("token", response.data.myNewToken);
-      const newAccessToken = response.data.myNewToken;
-      return instance({
-        ...response.config,
-        headers: {
-          Authorization: `${newAccessToken}`,
-        },
-      });
+      try {
+        window.localStorage.removeItem("token");
+        window.localStorage.setItem("token", data.myNewToken);
+        instance(
+          (prevRequest.headers = {
+            "Content-Type": "application/json",
+            Authorization: `${data.myNewToken}`,
+          })
+        );
+        window.location.reload();
+        return await instance(prevRequest);
+      } catch (err) {
+        console.log(err);
+      }
     }
+
     return response;
   },
   (error) => {
-    if (error.status === 401) {
+    console.log("나 인터셉터에러", error);
+    if (error.status === 401 || error.response.status === 412) {
       Swal.fire({
         text: "로그인 시간이 만료되었습니다. 다시 로그인해주세요!",
         width: "300px",
@@ -50,9 +55,12 @@ instance.interceptors.response.use(
         confirmButtonColor: "#40d295",
         showClass: { popup: "animated fadeInDown faster" },
         hideClass: { popup: "animated fadeOutUp faster" },
+      }).then((result) => {
+        if (result.isConfirmed) {
+          window.localStorage.clear();
+          window.location.replace("/login");
+        }
       });
-      window.localStorage.clear();
-      window.location.replace("/login");
     }
     return Promise.reject(error);
   }
@@ -62,8 +70,8 @@ instance.interceptors.response.use(
 export const LoginAPI = {
   login: (payload) => instance.post(`users/login`, payload),
   // 소셜로그인(카카오)
-  kakaoLogin: (payload) => instance.get(`kakao/code?code=${payload}`),
-  googleLogin: (payload) => instance.get(`google/code?code=${payload}`),
+  kakaoLogin: (payload) => instance.get(`auth/kakao?code=${payload}`),
+  googleLogin: (payload) => instance.get(`auth/google?code=${payload}`),
 
   // 인증번호
   postforFindIdPw: (payload) => instance.post(`users/sendSms`, payload), //-> 아이디비밀번호찾기시
@@ -71,7 +79,6 @@ export const LoginAPI = {
   postforCheckVCode: (payload) => instance.post(`users/checkSms`, payload),
   // 아이디 찾기
   findId: (payload) =>
-    // console.log(payload),
     instance.post(`users/findId`, {
       phone: payload.phone.phone,
       code: payload.code.code,
@@ -89,8 +96,7 @@ export const SignUpAPI = {
   signUp: (payload) => instance.post(`users/signup`, payload),
   checkId: (payload) => instance.post(`users/checkId`, payload),
   checkNickname: (payload) => instance.post(`/users/checkNick`, payload),
-  // checkPhoneNum: (payload) => instance.post(`/users/checkPhone`, payload),
-  kakaoSingUp: (payload) => instance.post(`kakao/signup`, payload),
+  socialSignUp: (payload) => instance.post(`auth/signup`, payload),
 };
 
 // userpage
@@ -105,8 +111,7 @@ export const UserpageAPI = {
       },
     }),
   deleteTeam: (payload) => instance.delete(`teams/${payload}`),
-  patchMyInfo: (payload) => console.log(payload),
-  // instance.patch(`users/me`, payload),
+  patchMyInfo: (payload) => instance.patch(`users/me`, payload),
   patchMyPhoto: (payload) => instance.patch(`users/profileImg`, payload),
   patchMyTeam: (payload) => instance.patch(`teams`, payload),
   dropOutMe: (payload) => instance.patch(`users/drop`, payload),
@@ -116,7 +121,6 @@ export const UserpageAPI = {
 // spotsdetail 실제 예약 서비스
 export const SpotsMatchApi = {
   postSpotsMatch: (payload) => instance.post(`reservations/register`, payload),
-
   getAllMatch: (payload) =>
     instance.get(`reservations/register/${payload.place}/${payload.date}`, {
       place: payload.place,
@@ -130,13 +134,13 @@ export const SpotsMatchApi = {
         date: payload.date,
       }
     ),
-
   getMyMatch: () => instance.get(`/reservations/me`),
   exitMyMatch: (payload) =>
     instance.put(`/reservations/register/delete`, payload),
   getRecentMatch: () => instance.get(`reservations/register`),
 };
 
+// 사설구장
 export const PrivateApi = {
   registerSpot: (payload) => instance.post(`places`, payload),
   // 등록한 사설 구장들 리스트
@@ -147,7 +151,6 @@ export const PrivateApi = {
   deletePrivateSpot: (payload) => instance.delete(`places/${payload}`),
   // 내가 등록한 구장 수정
   editPrivateSpot: (payload) =>
-    // console.log(payload),
     instance.patch(`/places/${payload.placesId}`, {
       spotName: payload.spotName,
       desc: payload.desc,
